requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
requests>=2.28.0
numpy>=1.23.0
sentence-transformers>=2.2.0
python-multipart>=0.0.5
pillow>=9.4.0
pytesseract>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.0